+++ tests/testthat/test-ci_deploy.R
@@ -1,14 +1,25 @@
 res <- restore_fixture()
 the_remote <- gert::git_remote_list(repo = res)
 remote_name <- "sandpaper-local"
-has_git <- Sys.which("git") != ""
+
+mask_output <- function(output, repo, remote) {
+  output <- gsub(repo, "[repo mask]", output, fixed = TRUE)
+  output <- gsub(remote, "[remote mask]", output, fixed = TRUE)
+  output <- gsub("[0-9a-f]{7,}", "[sha mask]", output)
+  no <- grepl("^Time", output)   | # no timestamps
+    grepl("^Author", output)     | # no author
+    grepl("^ ", output)            # no specific files
+  output[!no]
+}
 
 test_that("ci_deploy() will deploy once", {
 
-  skip_if_not(has_git)
+  skip_if_not(has_git())
   skip_if_not(rmarkdown::pandoc_available("2.11"))
 
-  ci_deploy(res, md_branch = "MD", site_branch = "SITE", remote = remote_name)
+  out1 <- capture.output({
+    ci_deploy(res, md_branch = "MD", site_branch = "SITE", remote = remote_name)
+  })
   expected <- expand.grid(
     c("refs/heads", "refs/remotes/sandpaper-local"),
     c("main", "MD", "SITE")
@@ -28,20 +39,45 @@ test_that("ci_deploy() will deploy once", {
 
 test_that("ci_deploy() will fetch sources from upstream", {
 
-  skip_if_not(has_git)
+  skip_if_not(has_git())
   skip_if_not(rmarkdown::pandoc_available("2.11"))
+  expect_true(gert::git_branch_exists("MD", local = TRUE, repo = res))
+  expect_true(gert::git_branch_exists("SITE", local = TRUE, repo = res))
+  gert::git_branch_delete("MD", repo = res)
+  gert::git_branch_delete("SITE", repo = res)
+  expect_false(gert::git_branch_exists("MD", local = TRUE, repo = res))
+  expect_false(gert::git_branch_exists("SITE", local = TRUE, repo = res))
   # The built directory does _not_ exist right now
   expect_false(fs::dir_exists(path_built(res)))
 
-  suppressMessages({expect_message(
+  out2 <- capture.output({suppressMessages({expect_message(
     ci_deploy(res, md_branch = "MD", site_branch = "SITE", remote = remote_name),
-    "nothing to commit!"
-  )})
-  md_log   <- gert::git_log("MD", repo = res)
-  site_log <- gert::git_log("SITE", repo = res)
-  expect_equal(nrow(md_log), 2)
-  skip("Behavior of the site worktree needs investigating")
-  expect_equal(nrow(site_log), 2)
+    "nothing to commit on MD!"
+  )})})
+  # the built directory is cleaned up afterwards
+  expect_false(fs::dir_exists(path_built(res)))
 
+  # The branches exist and nothing new has been committed to the MD branch.
+  expect_true(gert::git_branch_exists("MD", local = TRUE, repo = res))
+  expect_true(gert::git_branch_exists("SITE", local = TRUE, repo = res))
+  md_log   <- gert::git_log("MD", repo = res)
+  expect_equal(nrow(md_log), 2)
+
+})
+
+test_that("bundle_pr_artifacts() can record diffs", {
+
+  skip("still working on this test")
+
+  withr::with_dir(res, {
+    # built worktree
+    del_md <- git_worktree_setup(res, fs::path(res, "site", "built"), 
+      branch = "MD", remote = remote_name
+    )
+    # ------------ site worktree
+    del_site <- git_worktree_setup(res, fs::path(res, "site", "docs"),
+      branch = "SITE", remote = remote_name
+    )
+  })
 })